+++ src/lib/libheap/heap_block_size.c
@@ -1,7 +1,7 @@
 
 #include "_heap.h"
 
-int heap_block_size(heap_block_t* block)
+size_t heap_block_size(heap_block_t* block)
 {
     return *block < 0 ? -*block : *block;
 }
