+++ src/lib/libheap/heap_realloc.c
@@ -1,9 +1,9 @@
 
 #include "_heap.h"
 
-void* heap_realloc(heap_t* heap, void* memory, int newsize)
+void* heap_realloc(heap_t* heap, void* memory, size_t newsize)
 {
-    int oldsize = heap_msize(memory);
+    size_t oldsize = heap_msize(memory);
     if (oldsize != newsize) {
         void* newmem = heap_malloc(heap, newsize);
         if (newmem) {
